// shared/types/enums.ts
export enum ButtonTypes {
  PRIMARY = 'primary',
  SECONDARY = 'secondary',
  DEFAULT = 'default',
  LINK = 'link'
}

export enum Prices {
  STANDARD = 'standard',
  COMFORT = 'comfort',
  COMFORT_PLUS = 'comfort-plus',
  BUSINESS = 'business',
  MINIVAN = 'minivan',
  DELIVERY = 'delivery'
}

export enum Blocks {
  WELCOME = 'Первый блок',
  CALCULATOR = 'Калькулятор',
  ORDER_STEPS = '3 шага для заказа такси',
  QUESTIONS = 'Остались вопросы ?',
  TRANSFERS_EMPLOYEES_MANAGERS = 'Трансферы для сотрудников и руководителей',
  EDO_DYADOK = 'Работа по Эдо.Диадок',
  LEGAL_ENTITIES = 'Расчет для юр.лиц',
  FOOTER = 'Футер',
}

export enum WeatherType {
  SUNNY = 'SUNNY',
  CLOUDY = 'CLOUDY',
  PARTLY_CLOUDY = 'PARTLY_CLOUDY',
  SNOW = 'SNOW,',
  LITTLE_RAIN = 'LITTLE_RAIN',
  RAINFALL = 'RAINFALL',
  THUNDERSTORM = 'THUNDERSTORM',
  WINDY = 'WINDY'
}